requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings

psycopg
psycopg-binary
asyncpg
sqlalchemy_utc

pytest
pytz

sqladmin